update lab1-8 codes and docs. now version is 0.2

+++ code/lab8/user/libs/syscall.h
@@ -11,7 +11,7 @@ int sys_getpid(void);
 int sys_putc(int c);
 int sys_pgdir(void);
 int sys_sleep(unsigned int time);
-size_t sys_gettime(void);
+int sys_gettime(void);
 
 struct stat;
 struct dirent;
